+++ app/controllers/staff/activity_forms_controller.rb
@@ -25,6 +25,7 @@ class Staff::ActivityFormsController < Staff::BaseController
     :intended_beneficiaries,
     :flow,
     :aid_type,
+    :oda_eligibility,
   ]
 
   steps(*FORM_STEPS)
@@ -123,6 +124,8 @@ class Staff::ActivityFormsController < Staff::BaseController
       @activity.assign_attributes(flow: flow)
     when :aid_type
       @activity.assign_attributes(aid_type: aid_type)
+    when :oda_eligibility
+      @activity.assign_attributes(oda_eligibility: oda_eligibility)
     end
 
     update_form_state
@@ -234,6 +237,10 @@ class Staff::ActivityFormsController < Staff::BaseController
 
   def aid_type
     params.require(:activity).permit(:aid_type).fetch("aid_type", nil)
+  end
+
+  def oda_eligibility
+    params.require(:activity).permit(:oda_eligibility).fetch("oda_eligibility", nil)
   end
 
   def finish_wizard_path