+++ db/migrate/20200907094528_add_oda_eligibility_to_activities.rb
@@ -1,0 +1,5 @@
+class AddOdaEligibilityToActivities < ActiveRecord::Migration[6.0]
+  def change
+    add_column :activities, :oda_eligibility, :boolean, default: true
+  end
+end