+++ CHANGELOG.md
@@ -291,9 +291,9 @@
 - Show Level A (Fund), organisation and financial quarter on the report edit
   page
 - Handle attempts to activate invalid reports
-
 - `Recipient_region` codelist modified
 - `Intended beneficiaries` form field added, including validations, and country to region mapping
+- `ODA eligibility` form step added to create activity journey
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-15...HEAD
 [release-15]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-14...release-15